token-rotation/lab1-cicd-rotation/requirements.txt
# Python dependencies for GitLab Token Rotation Lab

# GitLab API client
python-gitlab>=4.0.0

# HashiCorp Vault client
hvac>=2.0.0

# Additional utilities
requests>=2.31.0
python-dateutil>=2.8.2
